vignan_chatbot/requirements.txt
fastapi
uvicorn
google-generativeai
jinja2
python-dotenv
